[poincare] Template for double

+++ apps/graph/cartesian_function.h
@@ -11,7 +11,7 @@ public:
   CartesianFunction(const char * text = nullptr, KDColor color = KDColorBlack);
   bool displayDerivative();
   void setDisplayDerivative(bool display);
-  float approximateDerivative(float x, Poincare::Context * context) const;
+  double approximateDerivative(double x, Poincare::Context * context) const;
   char symbol() const override;
 private:
   bool m_displayDerivative;
